Practice Mongo/fastapi/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-dotenv
pymongo